requirements.txt
# Основные пакеты
python-telegram-bot==20.8
python-dotenv==1.0.0
aiohttp==3.9.3
aiofiles==23.2.1

# Для работы с аудио и произношением
gTTS==2.5.1

# Для управления зависимостями
pip-tools==7.4.1 